Implement create profile functionality
Added a 'createProfile' function to the customerServices and used it in 'CreateProfile' page. A new type 'CreateProfileInputs' is also defined. The function allows for the creation of a new customer profile, including error handling for unsuccessful creations. Refactoring was done on the 'CreateProfile' page from handling the action itself to using the service function.